Refactor DocCursor out of models.
The DocNode model was getting a bit heavy with logic, all of which could be
shifted to the DocCursor (/tree). This moves that logic and the DocCursor out
to lighten the models file.
