+++ api/document/tests/utils.py
@@ -1,7 +1,8 @@
 import random
 from string import ascii_lowercase
 
-from document.models import DocCursor, DocNode
+from document.models import DocNode
+from document.tree import DocCursor
 
 
 def node_type(length: int = None) -> str:
@@ -13,7 +14,7 @@ def node_type(length: int = None) -> str:
 def random_doc(num_nodes: int = 10, save: bool = False, **kwargs) -> DocCursor:
     """Generates a random document with the requested number of nodes. This
     makes no guarantees about the shape of the resulting tree."""
-    root = DocNode.new_tree(node_type(), '0', **kwargs)
+    root = DocCursor.new_tree(node_type(), '0', **kwargs)
     node_count = num_nodes - 1  # subtract one for the root
     to_process = [root]
     while to_process:
