+++ api/document/views.py
@@ -3,6 +3,16 @@ from rest_framework.generics import RetrieveAPIView
 
 from document.models import DocNode
 from document.serializers import DocCursorSerializer
+from document.tree import DocCursor
+
+
+def optimize(queryset):
+    """To avoid the "n+1" query problem, we will optimize our querysets by
+    either joining 1-to-1 relations (via select_related) or ensuring a single
+    query for many-to-many relations (via prefetch_related)."""
+    return queryset.\
+        select_related('requirement').\
+        prefetch_related('footnotecitations', 'requirement__topics')
 
 
 class TreeView(RetrieveAPIView):
@@ -15,7 +25,7 @@ class TreeView(RetrieveAPIView):
             query_args['identifier'] = self.kwargs['identifier']
         else:
             query_args['depth'] = 0
-        queryset = DocNode.objects.select_related('requirement').\
-            prefetch_related('footnotecitations', 'requirement__topics')
-        root_struct = get_object_or_404(queryset, **query_args)
-        return root_struct.subtree(queryset)
+        root_doc = get_object_or_404(optimize(DocNode.objects), **query_args)
+        root = DocCursor.load_from_model(root_doc, subtree=False)
+        root.add_models(optimize(root_doc.descendants()))
+        return root
